% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_analyser.R
\name{analyse_long_range_links}
\alias{analyse_long_range_links}
\title{analyse_long_range_links}
\usage{
analyse_long_range_links(
  dset,
  lr_links_path,
  sr_links_path,
  are_lrlinks_ordered = F,
  SnpEff_Annotate = F,
  snpeff_jar_path = NULL,
  gbk_path = NULL,
  gff3_path = NULL,
  ref_fasta_path = NULL,
  validate_ref_ann_lengths = T,
  snp.dat = NULL,
  cds_var = NULL,
  max_tophits = 500,
  links_from_spydrpick = F,
  mega_dset = F
)
}
\arguments{
\item{dset}{name of the dataset, all outputs will be saved to the folder <dset>, possible to use the same dset name used for the short-range analysis}

\item{lr_links_path}{path to saved LR links, generally called <lr_links.tsv> or the SpydrPick edges (links) file}

\item{sr_links_path}{path to saved SR links, generally called <sr_links.tsv>}

\item{are_lrlinks_ordered}{are the links sorted in descending MI order (default = F), leave at default if unsure}

\item{SnpEff_Annotate}{specify whether to perform annotations using SnpEff (default = F)}

\item{snpeff_jar_path}{path to <snpEff.jar> (default = NULL). If annotations are not required, set SnpEff_Annotate = F}

\item{gbk_path}{path to genbank annotations file (default = NULL). Only provide one of genbank or gff3 annotation files.}

\item{gff3_path}{path to gff3 annotations file (default = NULL). Only provide one of genbank or gff3 annotation files.}

\item{ref_fasta_path}{path to Reference fasta file. The file MUST be in fasta format and contain exactly one sequence! Required for gff3 annotations,
not required for genbank annotations if the file contains the reference sequence.}

\item{validate_ref_ann_lengths}{check if the gbk reference sequence length matches the with fasta alignment (default = T)}

\item{snp.dat}{output from LDWeaver::parse_fasta_alignment(), not needed if SnpEff_Annotate = F}

\item{cds_var}{output from LDWeaver::estimate_variation_in_CDS(), not needed if SnpEff_Annotate = F}

\item{max_tophits}{specify the maximum number of long range links to save as <lr_tophits.tsv>. Note: all short-range links will be annotated (and saved separately),
but only the top <max_tophits> will be used for visualisation (default = 500)}

\item{links_from_spydrpick}{are the links computed using spydrpick (default = F)}

\item{mega_dset}{set TRUE for mega scale datasets (default = F)}
}
\description{
Function to analyse long range links. These links can be computed using LDWeaver or Spydrpick (faster). Spydrpick is recommended for larger datasets.
}
\examples{
\dontrun{
analyse_long_range_links = function("dset", "dset/lr_links.tsv", "dset/sr_links.tsv")
}
}
