% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BacGWES.R
\name{BacGWES}
\alias{BacGWES}
\title{BacGWES}
\usage{
BacGWES(
  dset,
  aln_path,
  gbk_path,
  check_gbk_fasta_lengths = T,
  snp_filt_method = "default",
  snpeff_jar_path = NULL,
  hdw_threshold = 0.1,
  SnpEff_Annotate = T,
  sr_dist = 20000,
  lr_retain_level = 0.99,
  max_tophits = 250,
  num_clusts_CDS = 3,
  srp_cutoff = 3,
  tanglegram_break_segments = 5,
  multicore = T,
  max_blk_sz = 10000,
  ncores = NULL
)
}
\arguments{
\item{dset}{name of the dataset, all outputs will be saved to the folder <dset>}

\item{aln_path}{path to the multi fasta alignment}

\item{gbk_path}{path to genbank file}

\item{check_gbk_fasta_lengths}{check if the gbk reference sequence length matches the with fasta alignment (default = T)}

\item{snp_filt_method}{specify the filtering method for SNP extraction: 'relaxed' or 'default' (default = 'default')}

\item{snpeff_jar_path}{path to <snpEff.jar>. If unavailable or not required, set SnpEff_Annotate = F}

\item{hdw_threshold}{Hamming distance similarity threshold (default = 0.1, i.e. 10\%) - add more}

\item{SnpEff_Annotate}{specify whether to perform annotations using SnpEff}

\item{sr_dist}{links less than <sr_dist> apart are considered 'short range' (default = 20000), range 1000 - 25000 bp.}

\item{lr_retain_level}{specify the long-range MI retaining percentile (default = 0.99) - in each block, only the top 1\% of lr MI links will be retained, range 0.1-0.999. WARNING! Low values will results in a HUGE lr_links.tsv file!}

\item{max_tophits}{specify the maximum number of short range links to save as <tophits.tsv>. Note: all short-range links will be annotated (and saved separately),
but only the top <max_tophits> will be used for visualisation (default = 250), range 50 - 1000}

\item{num_clusts_CDS}{parition to genome into num_clusts_CDS regions using k-means (default = 3) range 1 - 10}

\item{srp_cutoff}{specify the short-range -log10(p) cut-off value to discard short-range links before returning the data.frame. This setting has no impact on the
modelling since all links are used. However, setting a threshold > 2 will generally reduce the memory usage, plotting time (default = 3, i.e. corresponding to p = 0.001),
and run time for ARACNE. If all links are required to be returned, set to 0 (i.e. corresponding to p = 1), range 0 - 5}

\item{tanglegram_break_segments}{specify the number of genome segments to prepare - one tanglegram per segment (default = 5), range 1 - 10}

\item{multicore}{specify whether to use parallel processing (default = T)}

\item{max_blk_sz}{specify maximum block size for MI computation (default = 10000), larger sizes require more RAM, range 1000 - 100000}

\item{ncores}{specify the number of cores to use for parallel processing (default = NULL), will auto detect if NULL}
}
\value{
Numeric Value 0 if successful (all generated outputs will be saved)
}
\description{
Function to run the BacGWES pipeline
}
\examples{
\dontrun{
sr_links_red = BacGWES(dset = "efcm", aln_path = "<efcm_aln>", gbk_path = "<efcm.gbk>")
}
}
