% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{cleanup}
\alias{cleanup}
\title{cleanup}
\usage{
cleanup(dset, delete_after_moving = F)
}
\arguments{
\item{dset}{<dset> folder that requires cleanup.}

\item{delete_after_moving}{specify whether the organised files should be deleted from the original location (default = F).
If False, all organised files will be moved to a directory called OLD}
}
\description{
Function to clean up the files in <dset> folder, this will organise the saved files in to a more sensible folder structure.
If clean up was performed already, nothing will be changed.
}
\examples{
\dontrun{
gff3_path = "<path_to_gff3_annotation"
gff = read_GFF3_Annotation(gff3_path)
}
}
