% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{cleanup_support}
\alias{cleanup_support}
\title{cleanup_support}
\usage{
cleanup_support(files, fldr)
}
\arguments{
\item{files}{files to copy <from>}

\item{fldr}{folder to copy <to>}
}
\value{
parsed annotation file in list format
}
\description{
File operations for the cleanup function
}
\examples{
\dontrun{
gff3_path = "<path_to_gff3_annotation"
gff = read_GFF3_Annotation(gff3_path)
}
}
