% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetworkPlot.R
\name{create_network_for_gene}
\alias{create_network_for_gene}
\title{create_network_for_gene}
\usage{
create_network_for_gene(
  gene_name,
  sr_annotated_path = NULL,
  lr_annotated_path = NULL,
  drop_syXsy = T,
  drop_indirect = T,
  level = 1,
  separator = ":",
  min_links_to_include = 3
)
}
\arguments{
\item{gene_name}{case sensitive name of genome region (must match the SnpEff annotated name)}

\item{sr_annotated_path}{path to short-range annotated links tsv file (default = NULL)}

\item{lr_annotated_path}{path to long-range annotated links tsv file (default = NULL)}

\item{drop_syXsy}{specify whether to drop synonymous-synonymous links (default = T)}

\item{drop_indirect}{specify whether to drop ARACNE indirect links (default = T)}

\item{level}{set 1 for only nearest neighbours, set 2 for an additional layer (default = 1)}

\item{separator}{break pattern in annotation (default = ":", this is the default used by SnpEff)}

\item{min_links_to_include}{specify the minimum number of links required to retain the node in the plot (default = 2)}
}
\value{
data_frame to use as input for BacGWES::create_network()
}
\description{
Function to generate a network view of GWES hits for a given gene
}
\examples{
\dontrun{
create_network_for_gene(tophits, netplot_path)
}

}
