% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSummaryPlot.R
\name{.mat}
\alias{.mat}
\title{.mat}
\usage{
.mat(n, r)
}
\arguments{
\item{n}{number of dimensions}

\item{r}{reduction factor}
}
\value{
kernel reduction
}
\description{
.mat
}
\examples{
\dontrun{
.mat(10, 5)
}
}
