% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSummaryPlot.R
\name{.rescale01}
\alias{.rescale01}
\title{.rescale01}
\usage{
.rescale01(v)
}
\arguments{
\item{v}{vector to rescale between 0-1}
}
\value{
rescaled vector
}
\description{
.rescale01
}
\examples{
\dontrun{
.rescale01(sample(10,25,10))
}
}
