% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performPopulationStuctureCorrection.R
\name{estimate_Hamming_distance_weights}
\alias{estimate_Hamming_distance_weights}
\title{estimate_Hamming_distance_weights}
\usage{
estimate_Hamming_distance_weights(snp.dat, threshold = 0.1)
}
\arguments{
\item{snp.dat}{output from parsing the multi fasta alignment using BacGWES::parse_fasta_alignment()}

\item{threshold}{Hamming distance similarity threshold (default = 0.1, i.e. 10\%)}
}
\value{
Vector with Hamming distance weights for each sequence that can be used to reweight Mutual Information estimates
}
\description{
Function to estimate the similarity between sequences based on the Hamming distance
}
\examples{
\dontrun{
hdw <- estimate_Hamming_distance_weights(snp.dat, threshold=0.1)
}
}
