% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCDSDiversity.R
\name{estimate_variation_in_CDS}
\alias{estimate_variation_in_CDS}
\title{estimate_variation_in_CDS}
\usage{
estimate_variation_in_CDS(gbk, snp.dat, ncores)
}
\arguments{
\item{gbk}{output from parsing the genbank file using BacGWES::parseGenBankFile()}

\item{snp.dat}{output from parsing the multi fasta alignment using BacGWES::parseFastaAlignment()}

\item{cores}{specify the number of cores to use}
}
\value{
R list with CDS variation and allele distribution details
}
\description{
Function to estimate the variation within each coding region, the output from this function
can be used to segment the genome into diversity-based clusters.
}
\examples{
\dontrun{
snp.dat <- estimate_variation_in_CDS(gbk, snp.dat, ncores=10)
}
}
