% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{generate_Links_SNPS_fasta}
\alias{generate_Links_SNPS_fasta}
\title{generate_Links_SNPS_fasta}
\usage{
generate_Links_SNPS_fasta(
  snp.dat,
  aln_path,
  pos_path,
  lr_tophits_path = NULL,
  lr_annotated_links_path = NULL,
  sr_tophits_path = NULL,
  sr_annotated_links_path = NULL
)
}
\arguments{
\item{snp.dat}{output from parsing the multi fasta alignment using LDWeaver::parse_fasta_alignment(), generally saved to Additional_Outputs/snp_ACGTN.rds}

\item{aln_path}{path to save output fasta file or tsv file}

\item{pos_path}{path to write the SNP positions, only required for format = "fasta" (default = NULL)}

\item{lr_tophits_path}{Path to the lr_tophits file (typically Tophits/lr_tophits.tsv).}

\item{lr_annotated_links_path}{Path to the lr_annotated_links file (typically Annotated_links/lr_links_annotated.tsv).}

\item{sr_tophits_path}{Path to the lr_tophits file (typically Tophits/sr_tophits.tsv).}

\item{sr_annotated_links_path}{Path to the sr_annotated_links file (typically Annotated_links/sr_links_annotated.tsv).}

\item{pos}{set of loci to be included in the saved alignment (default = NULL - save everything). <pos> should be a numeric vector comprising a subset of sites in snp.dat$POS and cannot contain duplicates}

\item{format}{Output can be presented as a standard fasta file, or an optional tsv (default = "fasta")}
}
\description{
Create a link SNP only fasta file. This file is required to generate (detailed) tree plots
}
