% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGWESplots.R
\name{make_gwes_plots}
\alias{make_gwes_plots}
\title{make_gwes_plots}
\usage{
make_gwes_plots(lr_links = NULL, sr_links = NULL, plt_folder = NULL)
}
\arguments{
\item{lr_links}{data.frame containing lr_links or the path to saved lr_links TSV file.
Usually output from perform_MI_computation(), (default = NULL)}

\item{sr_links}{data.frame containing sr_links or the path to saved sr_links TSV file.
Usually output from perform_MI_computation(), (default = NULL)}

\item{plt_folder}{specify the folder to save generated plots (default = NULL, will be saved to a folder called PLOTS in getwd())}
}
\value{
R data frame with short range GWES links (plots and long range links will be saved)
}
\description{
Function to generate long and short range GWES plots
}
\examples{
\dontrun{
sr_links_red = perform_MI_computation(snp.dat = snp.dat, hdw = hdw, cds_var = cds_var, ncores = 10,
lr_save_path = "lr.tsv", plt_folder = "plots")
}

}
