% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSNPs_mega.R
\name{parse_fasta_alignment_mega}
\alias{parse_fasta_alignment_mega}
\title{extractSNPs_from_megaMSA}
\usage{
parse_fasta_alignment_mega(
  aln_path,
  gap_freq = 0.15,
  maf_freq = 0.01,
  method = "default"
)
}
\arguments{
\item{aln_path}{path to multi fasta alignment}

\item{gap_freq}{sites with a gap frequency >gap_greq will be dropped (default = 0.15)}

\item{maf_freq}{sites with a minor allele frequency <maf_freq will be dropped (default = 0.01)}

\item{method}{specify the filtering method 'relaxed' or 'default' (default = 'default')}
}
\value{
R list with SNPs in sparse format and additional parameters
}
\description{
Function to extract SNPs from mega scale multi fasta-alignments.
Depending on alignment size, this function can take a very long time. Best to run snp-sites externally and use the snp-only alignment with parse_fasta_SNP_alignment_mega)
}
\examples{
\dontrun{
aln_path <- system.file("extdata", "sample.aln", package = "LDWeaver")
snp.dat <- parseFastaAlignment(aln_path, gap_freq = 0.15, maf_freq = 0.01, method = "default")
}
}
