% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGFF.R
\name{parse_gff_file}
\alias{parse_gff_file}
\title{parse_gff_file}
\usage{
parse_gff_file(gff3_path, ref_fasta_path, perform_length_check = T)
}
\arguments{
\item{gff3_path}{path to gff3_annotation file. This function relies on ape::read.gff() to read in the file}

\item{ref_fasta_path}{path to Reference fasta file. The file MUST be in fasta format and contain exactly one sequence!}

\item{perform_length_check}{specify whether to check if the gff3 annotations agree with the reference length (default = T)}
}
\value{
gff annotation and reference sequence as a list object
}
\description{
Function to parse the gff3 file for the fasta alignment.
}
\examples{
\dontrun{
gff3_path <- "<path_to_gff3_file>"
ref_fasta_path <- "<path_to_reference_fasta>"
ref_gff <- parse_gff_file(gff3_path, ref_fasta_path)
}
}
