% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePairwiseMI.R
\name{perform_MI_computation}
\alias{perform_MI_computation}
\title{perform_MI_computation}
\usage{
perform_MI_computation(
  snp.dat,
  hdw,
  cds_var,
  ncores,
  lr_save_path = NULL,
  sr_save_path = NULL,
  plt_folder = NULL,
  sr_dist = 20000,
  lr_retain_level = 0.99,
  max_blk_sz = 10000,
  srp_cutoff = 3,
  runARACNE = TRUE,
  perform_SR_analysis_only = FALSE,
  order_links = T
)
}
\arguments{
\item{snp.dat}{output from parsing the multi fasta alignment using BacGWES::parse_fasta_alignment()}

\item{hdw}{vector of Hamming distance weights, output from BacGWES::estimate_Hamming_distance_weights()}

\item{cds_var}{output from BacGWES::estimate_variation_in_CDS()}

\item{ncores}{specify the number of cores to use}

\item{lr_save_path}{specify the location to save long range MI links as a tsv file (default = NULL, will be auto set)}

\item{sr_save_path}{specify the location to save short range MI links as a tsv file (default = NULL, will be auto set), links below srp_cutoff will not be saved}

\item{plt_folder}{specify the folder to save generated plots (default = NULL, will be saved to a folder called PLOTS in getwd())}

\item{sr_dist}{specify the short-range basepair separation (default = 20000)}

\item{lr_retain_level}{specify the long-range MI retaining percentile (default = 0.99) - in each block, only the top 1\% of lr MI links will be retained}

\item{max_blk_sz}{specify maximum block size for MI computation (default = 10000), larger sizes require more RAM}

\item{srp_cutoff}{specify the short-range -log10(p) cut-off value to discard short-range links before returning the data.frame. This setting has no impact on the
modelling since all links are used. However, setting a threshold > 2 will generally reduce the memory usage, plotting time (default = 3, i.e. corresponding to p = 0.001),
and run time for ARACNE. If all links are required to be returned, set to 0 (i.e. corresponding to p = 1)}

\item{runARACNE}{specify whether to run ARACNE (default = TRUE), if set to FAULT, all links will be marked as ARACNE=1}

\item{perform_SR_analysis_only}{skip the long range link analysis (default = FALSE)}

\item{order_links}{return and save links after sorting in short-range p-value order, most to least important (default = T)}
}
\value{
R data frame with short range GWES links (plots and long range links will be saved)
}
\description{
Function to compute pairwise MI (long range links saved to text file), model
the background LD levels within each cluster (plots saved) and
compute short-range p-values for each short-range links. If requested, this
function can also run the ARACNE algorithm
}
\examples{
\dontrun{
sr_links_red = perform_MI_computation(snp.dat = snp.dat, hdw = hdw, cds_var = cds_var, ncores = 10,
lr_save_path = "lr.tsv", plt_folder = "plots")
}

}
