% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_AnnotatedLinks}
\alias{read_AnnotatedLinks}
\title{read_AnnotatedLinks}
\usage{
read_AnnotatedLinks(annotated_links_path)
}
\arguments{
\item{annotated_links_path}{path to saved <sr> or <lr>_annotated_links.tsv file}
}
\value{
Annotated links as a data.frame
}
\description{
read_AnnotatedLinks
}
\examples{
\dontrun{
sr_annotated_links_path = "<path_to_annotated_sr_links.tsv_file>"
sr_links_ann = read_AnnotatedLinks(sr_annotated_links_path)
}
}
