% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_GFF3_Annotation}
\alias{read_GFF3_Annotation}
\title{read_GFF3_Annotation}
\usage{
read_GFF3_Annotation(gff3_path)
}
\arguments{
\item{gff3_path}{path to gff3_annotation file. This function relies on ape::read.gff() to read in the file,
refer to ?ape::read.gff for additional help}
}
\value{
parsed annotation file in list format
}
\description{
read_GFF3_Annotation
}
\examples{
\dontrun{
gff3_path = "<path_to_gff3_annotation"
gff = read_GFF3_Annotation(gff3_path)
}
}
