% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_ReferenceFasta}
\alias{read_ReferenceFasta}
\title{read_ReferenceFasta}
\usage{
read_ReferenceFasta(ref_fasta_path)
}
\arguments{
\item{ref_fasta_path}{path to Reference fasta file. The file MUST be in fasta format and contain exactly one sequence!}
}
\value{
parsed reference sequence with metadata in list format
}
\description{
read_ReferenceFasta
}
\examples{
\dontrun{
ref_fa = "<path_to_ref_fasta"
reference = read_ReferenceFasta(ref_fa)
}
}
