% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_TopHits}
\alias{read_TopHits}
\title{read_TopHits}
\usage{
read_TopHits(top_hits_path)
}
\arguments{
\item{top_hits_path}{path to saved top_hits file}
}
\value{
Top hits as a data.frame
}
\description{
read_TopHits
}
\examples{
\dontrun{
top_hits_path = "<path_to_tophits_file>"
tophits = read_TopHits(top_hits_path)
}
}
