% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{readLongRangeLinks}
\alias{readLongRangeLinks}
\title{readLongRangeLinks}
\usage{
readLongRangeLinks(lr_links_path, links_from_spydrpick = F, sr_dist = 20000)
}
\arguments{
\item{lr_links_path}{path to saved lr_links.tsv or spydrpick file}

\item{links_from_spydrpick}{are the links computed using spydrpick (default = F)}

\item{sr_dist}{links less than <sr_dist> apart will be dropped from the data.frame (default = 20000)}
}
\value{
Long range links as a data.frame
}
\description{
readLongRangeLinks
}
\examples{
\dontrun{
lr_links_path = "<path_to_lrlinks.tsv_file>"
lr_links = readLongRangeLinks(lr_links_path)
}
}
