% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{readShortRangeLinks}
\alias{readShortRangeLinks}
\title{readShortRangeLinks}
\usage{
readShortRangeLinks(sr_links_path)
}
\arguments{
\item{sr_links_path}{path to saved sr_links.tsv file}
}
\value{
Long range links as a data.frame
}
\description{
readShortRangeLinks
}
\examples{
\dontrun{
sr_links_path = "<path_to_sr_links.tsv_file>"
sr_links = readShortRangeLinks(sr_links_path)
}
}
