% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{readTopHits}
\alias{readTopHits}
\title{readTopHits}
\usage{
readTopHits(top_hits_path)
}
\arguments{
\item{top_hits_path}{path to saved top_hits file}
}
\value{
Top hits as a data.frame
}
\description{
readTopHits
}
\examples{
\dontrun{
top_hits_path = "<path_to_tophits_file>"
tophits = readTopHits(top_hits_path)
}
}
