% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{runARACNE}
\alias{runARACNE}
\title{runARACNE}
\usage{
runARACNE(links_to_check, links_full)
}
\arguments{
\item{links_to_check}{data.frame comprising subset of links to run ARACNE on}

\item{links_full}{data.frame with all links}
}
\value{
Long range links as a data.frame
}
\description{
runARACNE
}
\examples{
\dontrun{
sr_links_path = "<path_to_sr_links.tsv_file>"
sr_links = readShortRangeLinks(sr_links_path)
sr_links_red = sr_links[which(sr_links$srp > 3), ]
sr_links_red$ARACNE = runARACNE(sr_links_red, sr_links)
}
}
