% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{snpdat_to_fa}
\alias{snpdat_to_fa}
\title{snpdat_to_fa}
\usage{
snpdat_to_fa(snp.dat, aln_path, pos_path = NULL, pos = NULL, format = "fasta")
}
\arguments{
\item{snp.dat}{output from parsing the multi fasta alignment using LDWeaver::parse_fasta_alignment()}

\item{aln_path}{path to save output fasta file or tsv file}

\item{pos_path}{path to write the SNP positions, only required for format = "fasta" (default = NULL)}

\item{pos}{set of loci to be included in the saved alignment (default = NULL - save everything). <pos> should be a numeric vector comprising a subset of sites in snp.dat$POS and cannot contain duplicates}

\item{format}{Output can be presented as a standard fasta file, or an optional tsv (default = "fasta")}
}
\description{
Create a SNP only fasta or tsv file from snp.dat. The user has the option to include only chosen positions.
}
