% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preptrees.R
\name{view_tree}
\alias{view_tree}
\title{view_tree}
\usage{
view_tree(
  tree_path,
  perform_midpoint_rooting = T,
  metadata_df = NULL,
  fasta_path,
  pos_file_path,
  links_df = NULL,
  lr_tophits_path = NULL,
  lr_annotated_links_path = NULL,
  sr_tophits_path = NULL,
  sr_annotated_links_path = NULL,
  ntop_links = 10,
  from = NULL,
  to = NULL,
  offset_metadata = NULL,
  offset_alleles = NULL,
  width_metadata = NULL,
  width_alleles = NULL,
  plot_save_path = NULL,
  plot_height = 20,
  plot_width = 15
)
}
\arguments{
\item{tree_path}{Path to tree file (must be readable using ape::read.tree). tip.labels must exactly match the sequence names in the fasta and metadata files.}

\item{perform_midpoint_rooting}{Specify wheter the tree should be midpoint rooted (default = T)}

\item{metadata_df}{data.frame with metadata. Must contain a column labelled 'id' that exactly matches the sequence names in the fasta file and tip.labels of the tree.}

\item{fasta_path}{fasta file comprising the sites associated with links}

\item{pos_file_path}{position file containing fasta sites}

\item{links_df}{A data.frame comprising a subset of tophits links, if provided, links are not read from the files. NOTE: Each links must (at a minimum) contain the fields: <pos1> & <pos2>.
If alleles from the <ntop_links> are to be visualised, ensure that the links_df is sorted such that the most important link is at the top.}

\item{lr_tophits_path}{Path to the lr_tophits file (typically Tophits/lr_tophits.tsv).}

\item{lr_annotated_links_path}{Path to the lr_annotated_links file (typically Annotated_links/lr_links_annotated.tsv).}

\item{sr_tophits_path}{Path to the lr_tophits file (typically Tophits/sr_tophits.tsv).}

\item{sr_annotated_links_path}{Path to the sr_annotated_links file (typically Annotated_links/sr_links_annotated.tsv).}

\item{ntop_links}{Specify the number of links to plot from tophits file(s). ntop_links will be selected from each file if both LR and SR links should be visualised (default = 10).}

\item{from}{draw the SNPs that are within the region starting <from> and <to> below (default = NULL). SNPs that are also in links with SNPs from the chosen region will be plotted.}

\item{to}{If <from> is provided, <to> muts also be provided and <ntop_links> will be ignored (default = NULL).}

\item{offset_metadata}{specify the separation between tree and metadata layer. These parameters should be modified to tune the plot dimensions as required. (default = NULL)}

\item{offset_alleles}{specify the speciation between tree and alleles layer (default = NULL)}

\item{width_metadata}{specify the width of the metadata layer as a proportion of the tree size (default = NULL)}

\item{width_alleles}{specify the width of the alleles layer as a proportion of the tree size (default = NULL)}

\item{plot_save_path}{path to save the tree plot (default = NULL, show plot instead of saving)}

\item{plot_height}{if saving, specify the plot height in inches (default = 20)}

\item{plot_width}{if saving, specify the plot height in inches (default = 15). Modify these numbers to tune the plot dimensions as required.}
}
\description{
Function to visualise allele distributions and metadata aligned below the tree. The tree file must be readable with ape::read.tree() and <tip.labels> must match the sequence names in the
provided fasta file(s). Besides the tree file, this function requires minimum three inputs: (1) a links file or data.frame, (2) a SNP fasta alignment and (3) a position <text> file with each row containing
the genomic positions of the fasta alignment. The links file is straightforward: it can be either one of <Tophits> or the <Annotated_links> files, both short and long range links
can be provided. A subset of manually chosen links can also be provided as a data.frame. The fasta/position files can be generated using LDWeaver::generate_Links_SNPS_fasta().
If metadata is available, they can be visualised along with the allele data in a separate panel.
}
\examples{
\dontrun{
genomewide_LDMap(lr_links_path, sr_links_path, plot_save_path)
}

}
