% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGWESExplorerOutput.R
\name{write_output_for_gwes_explorer}
\alias{write_output_for_gwes_explorer}
\title{write_output_for_gwes_explorer}
\usage{
write_output_for_gwes_explorer(snp.dat, srlinks_tophits, gwes_explorer_folder)
}
\arguments{
\item{snp.dat}{output from parsing the multi fasta alignment using BacGWES::parse_fasta_alignment()}

\item{srlinks_tophits}{data frame with top short range GWES links, returned from BacGWES::perform_snpEff_annotations()}

\item{gwes_explorer_folder}{folder to save the inputs for GWESExplorer}
}
\value{
none
}
\description{
Function to generate long and short range GWES plots
}
\examples{
\dontrun{
write_output_for_gwes_explorer(snp.dat, srlinks_tophits, gwes_explorer_folder)
}

}
