% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_analyser.R
\name{analyse_long_range_links}
\alias{analyse_long_range_links}
\title{analyse_long_range_links}
\usage{
analyse_long_range_links(
  dset,
  lr_links_path,
  sr_links_path,
  are_lrlinks_ordered = F,
  SnpEff_Annotate = F,
  snpeff_jar_path = NULL,
  gbk_path = NULL,
  snp.dat = NULL,
  cds_var = NULL,
  max_tophits = 500,
  links_from_spydrpick = F
)
}
\arguments{
\item{dset}{name of the dataset, all outputs will be saved to the folder <dset>, possible to use the same dset name used for the short-range analysis}

\item{lr_links_path}{path to saved LR links, generally called <lr_links.tsv> or the SpydrPick links file}

\item{sr_links_path}{path to saved SR links, generally called <sr_links.tsv>}

\item{are_lrlinks_ordered}{are the links sorted in descending MI order (default = F), leave at default if unsure}

\item{SnpEff_Annotate}{specify whether to perform annotations using SnpEff (default = F)}

\item{snpeff_jar_path}{path to <snpEff.jar> (default = NULL). If annotations are not required, set SnpEff_Annotate = F}

\item{gbk_path}{path to genbank file, not needed if SnpEff_Annotate = F}

\item{snp.dat}{output from BacGWES::parse_fasta_alignment(), not needed if SnpEff_Annotate = F}

\item{cds_var}{output from BacGWES::estimate_variation_in_CDS(), not needed if SnpEff_Annotate = F}

\item{max_tophits}{specify the maximum number of long range links to save as <lr_tophits.tsv>. Note: all short-range links will be annotated (and saved separately),
but only the top <max_tophits> will be used for visualisation (default = 500)}

\item{links_from_spydrpick}{are the links computed using spydrpick (default = F)}
}
\description{
Function to analyse long range links. These links can be computed using BacGWES or Spydrpick (faster). Spydrpick is recommended for larger datasets.
}
\examples{
\dontrun{
analyse_long_range_links = function("dset", "dset/lr_links.tsv", "dset/sr_links.tsv")
}
}
