% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetworkPlot.R
\name{create_network}
\alias{create_network}
\title{create_network}
\usage{
create_network(
  srlinks_tophits,
  netplot_path,
  plot_title = NULL,
  separator = ":",
  max_plot_nodes = NULL,
  plot_w = 6000,
  plot_h = 4000
)
}
\arguments{
\item{srlinks_tophits}{data frame with top short range GWES links, returned from BacGWES::perform_snpEff_annotations()}

\item{netplot_path}{folder to save the network}

\item{plot_title}{title for the network plot (default = NULL)}

\item{separator}{break pattern in annotation (default = ":", this is the default used by SnpEff)}

\item{max_plot_nodes}{specify the maximum number of nodes to visualise in the plot (default = NULL, use all). Large values will result in a cluttered output, consider increasing plot size.}

\item{plot_w}{specify plot width in pixels (default = 6000)}

\item{plot_h}{specify plot height in pixels (default = 4000)}
}
\value{
none
}
\description{
Function to generate a network view of top GWES hits
}
\examples{
\dontrun{
create_network(srlinks_tophits, netplot_path)
}

}
