% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTanglegram.R
\name{create_tanglegram}
\alias{create_tanglegram}
\title{create_tanglegram}
\usage{
create_tanglegram(srlinks_tophits, gbk, tanglegram_folder, break_segments = 5)
}
\arguments{
\item{srlinks_tophits}{data frame with top short range GWES links, returned from BacGWES::perform_snpEff_annotations()}

\item{gbk}{output from parsing the genbank file using BacGWES::parse_genbank_file()}

\item{tanglegram_folder}{folder to save tanglegram(s)}

\item{break_segments}{specify the number of genome segments to prepare - one tanglegram per segment (default = 5)}
}
\value{
none
}
\description{
Function to generate long and short range GWES plots
}
\examples{
\dontrun{
create_tanglegram(srlinks_tophits, gbk, tanglegram_path, break_segments = 5)
}

}
