% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSummaryPlot.R
\name{._legendfun2}
\alias{._legendfun2}
\title{._legendfun2}
\usage{
._legendfun2(
  legend = c("Group A", "Group B"),
  lwd = 3,
  cex = 1.1,
  col = c("red", "blue"),
  ...
)
}
\arguments{
\item{legend}{legend}

\item{lwd}{lwd}

\item{cex}{cex}

\item{col}{col}

\item{...}{...}
}
\description{
._legendfun2
}
