% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCDSDiversity.R
\name{estimate_variation_in_CDS}
\alias{estimate_variation_in_CDS}
\title{estimate_variation_in_CDS}
\usage{
estimate_variation_in_CDS(
  snp.dat,
  ncores,
  gbk = NULL,
  gff = NULL,
  num_clusts_CDS = 3,
  clust_plt_path = NULL
)
}
\arguments{
\item{snp.dat}{output from parsing the multi fasta alignment using LDWeaver::parse_fasta_alignment()}

\item{ncores}{specify the number of cores to use}

\item{gbk}{output from parsing the genbank file using LDWeaver::parse_genbank_file()}

\item{gff}{output from parsing the gff3 file using LDWeaver::parse_gff_file()}

\item{num_clusts_CDS}{parition to genome into num_clusts_CDS regions using k-means (default = 3)}

\item{clust_plt_path}{specify path to save CDS variation plot}
}
\value{
R list with CDS variation and allele distribution details
}
\description{
Function to estimate the variation within each coding region, the output from this function
can be used to segment the genome into diversity-based clusters.
}
\examples{
\dontrun{
cds_var <- estimate_variation_in_CDS(gbk, snp.dat, ncores = 10)
}
}
