% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCDSDiversity.R
\name{estimate_variation_in_CDS}
\alias{estimate_variation_in_CDS}
\title{estimate_variation_in_CDS}
\usage{
estimate_variation_in_CDS(gbk, snp.dat, num_clusts_CDS = 3, ncores)
}
\arguments{
\item{gbk}{output from parsing the genbank file using BacGWES::parseGenBankFile()}

\item{snp.dat}{output from parsing the multi fasta alignment using BacGWES::parseFastaAlignment()}

\item{num_clusts_CDS}{parition to genome into num_clusts_CDS regions using k-means (default = 3)}

\item{ncores}{specify the number of cores to use}
}
\value{
R list with CDS variation and allele distribution details
}
\description{
Function to estimate the variation within each coding region, the output from this function
can be used to segment the genome into diversity-based clusters.
}
\examples{
\dontrun{
cds_var <- estimate_variation_in_CDS(gbk, snp.dat, num_clusts_CDS = 3, ncores=10)
}
}
