% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSummaryPlot.R
\name{genomewide_LDMap}
\alias{genomewide_LDMap}
\title{genomewide_LDMap}
\usage{
genomewide_LDMap(
  lr_links_path,
  sr_links_path,
  plot_save_path,
  plot_title = NULL,
  reducer = NULL,
  links_from_spydrpick = F
)
}
\arguments{
\item{lr_links_path}{path to saved <lr_links.tsv> or spydrpick edges (links) file}

\item{sr_links_path}{path to saved SR links, generally called <sr_links.tsv>}

\item{plot_save_path}{path to save the LD Map}

\item{plot_title}{title for the LD Map (default = NULL)}

\item{reducer}{positive value used to reduce the LDMap size (default = NULL, will use length(positions)/1e3).
The default paramter should work well, reducing the parameter will increase resolution at the cost of resources.}

\item{links_from_spydrpick}{are the links computed using spydrpick (default = F)}
}
\description{
Function to generate the bird's-eye view of the genomewide LD structure (measured as weighted-MI)
}
\examples{
\dontrun{
genomewide_LDMap(lr_links_path, sr_links_path, plot_save_path)
}

}
