% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGWESplots.R
\name{make_gwes_plots}
\alias{make_gwes_plots}
\title{make_gwes_plots}
\usage{
make_gwes_plots(
  lr_links = NULL,
  sr_links = NULL,
  plt_folder = NULL,
  are_srlinks_ordered = F
)
}
\arguments{
\item{lr_links}{data.frame containing lr_links or the path to saved lr_links TSV file.
Usually output from perform_MI_computation(), (default = NULL). Instead of using this function to generate the lr-gwes plot, use the dedicated
LDWeaver::analyse_long_range_links() function for long range link analysis (and plotting).}

\item{sr_links}{data.frame containing sr_links or the path to saved sr_links TSV file.
Usually output from perform_MI_computation(), (default = NULL)}

\item{plt_folder}{specify the folder to save generated plots (default = NULL, will be saved to a folder called PLOTS in getwd())}

\item{are_srlinks_ordered}{if False, links will be ordered before plotting, which will plot higher srp links on top of lower ones (default = F).
Set True only if the links are already in the preferred order. Note: This function will reverse the link order to match the ggplot's top to bottom plotting order.}
}
\description{
Function to generate long and short range GWES plots
}
\examples{
\dontrun{
sr_links_red = perform_MI_computation(snp.dat = snp.dat, hdw = hdw, cds_var = cds_var, ncores = 10,
lr_save_path = "lr.tsv", plt_folder = "plots")
}

}
