% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnpEffAnnotations.R
\name{perform_snpEff_annotations}
\alias{perform_snpEff_annotations}
\title{perform_snpEff_annotations}
\usage{
perform_snpEff_annotations(
  dset_name,
  annotation_folder,
  snpeff_jar,
  gbk,
  gbk_path,
  cds_var,
  sr_links,
<<<<<<< HEAD
  snp.dat,
  max_tophits = 250
)
}
\arguments{
\item{dset_name}{name of the dataset}

\item{annotation_folder}{folder to save annotatations}

\item{snpeff_jar}{path to <snpEff.jar>}

\item{gbk}{output from parsing the genbank file using BacGWES::parse_genbank_file()}

\item{gbk_path}{path to genbank file}

\item{cds_var}{output from BacGWES::estimate_variation_in_CDS()}

\item{sr_links}{data.frame containing sr_links output from perform_MI_computation()}

\item{snp.dat}{output from parsing the multi fasta alignment using BacGWES::parse_fasta_alignment()}

\item{max_tophits}{specify the number of top hits to return (all links will be annotated and saved to the annotations_folder)}
}
\value{
R data frame with the top short range GWES links with snpEff annotated embedded
=======
  snp.dat
)
}
\arguments{
\item{sr_links}{data.frame containing sr_links or the path to saved sr_links TSV file. Usually output from perform_MI_computation()}

\item{plt_folder}{specify the folder to save generated plots (default = NULL, will be saved to a folder called PLOTS in getwd())}

\item{lr_links}{data.frame containing lr_links or the path to saved lr_links TSV file. Usually output from perform_MI_computation()}
}
\value{
R data frame with short range GWES links (plots and long range links will be saved)
>>>>>>> 90e835452b69a8447074fc3659161cfb69a215b8
}
\description{
Function to generate long and short range GWES plots
}
\examples{
\dontrun{
<<<<<<< HEAD
sr_links_red = perform_snpEff_annotations(dset_name, annotation_folder, snpeff_jar, gbk,
gbk_path, cds_var, sr_links, snp.dat, max_tophits = 250)
=======
sr_links_red = perform_MI_computation(snp.dat = snp.dat, hdw = hdw, cds_var = cds_var, ncores = 10,
lr_save_path = "lr.tsv", plt_folder = "plots")
>>>>>>> 90e835452b69a8447074fc3659161cfb69a215b8
}

}
