% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnpEffAnnotations.R
\name{perform_snpEff_annotations}
\alias{perform_snpEff_annotations}
\title{perform_snpEff_annotations}
\usage{
perform_snpEff_annotations(
  dset_name,
  annotation_folder,
  snpeff_jar,
  gbk,
  gbk_path,
  cds_var,
  sr_links,
  snp.dat,
  tophits_path = NULL,
  max_tophits = 250
)
}
\arguments{
\item{dset_name}{name of the dataset}

\item{annotation_folder}{folder to save annotatations}

\item{snpeff_jar}{path to <snpEff.jar>}

\item{gbk}{output from parsing the genbank file using BacGWES::parse_genbank_file()}

\item{gbk_path}{path to genbank file}

\item{cds_var}{output from BacGWES::estimate_variation_in_CDS()}

\item{sr_links}{data.frame containing sr_links (reduced links can be used to save time) output from perform_MI_computation()}

\item{snp.dat}{output from parsing the multi fasta alignment using BacGWES::parse_fasta_alignment()}

\item{tophits_path}{specify the path to save tophit links (default = NULL). If NULL, will be saved to annotations_folder/tophits.tsv}

\item{max_tophits}{specify the number of top hits to return (all links will be annotated and saved to the annotations_folder)}
}
\value{
R data frame with the top short range GWES links with snpEff annotated embedded
}
\description{
Function to generate long and short range GWES plots
}
\examples{
\dontrun{
sr_links_red = perform_snpEff_annotations(dset_name, annotation_folder, snpeff_jar, gbk,
gbk_path, cds_var, sr_links, snp.dat, max_tophits = 250)
}

}
