% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preptrees.R
\name{view_tree}
\alias{view_tree}
\title{view_tree}
\usage{
view_tree(
  tree_path,
  perform_midpoint_rooting = T,
  metadata_df = NULL,
  lr_tophits_path = NULL,
  lr_fasta_path = NULL,
  lr_pos_path = NULL,
  sr_tophits_path = NULL,
  sr_fasta_path = NULL,
  sr_pos_path = NULL,
  ntop_links = 10,
  from = NULL,
  to = NULL,
  offset_metadata = NULL,
  offset_alleles = NULL,
  width_metadata = NULL,
  width_alleles = NULL,
  plot_save_path = NULL,
  plot_height = 20,
  plot_width = 15
)
}
\arguments{
\item{tree_path}{Path to tree file (must be readable using ape::read.tree). tip.labels must exactly match the sequence names in the fasta and metadata files.}

\item{perform_midpoint_rooting}{Specify wheter the tree should be midpoint rooted (default = T)}

\item{metadata_df}{data.frame with metadata. Must contain a column labelled 'id' that exactly matches the sequence names in the fasta file and tip.labels of the tree.}

\item{lr_tophits_path}{Path to Tophits/lr_tophits.tsv file.If LR link tophits should be visualised, the two lr_files (below) must also be provided here.}

\item{lr_fasta_path}{Path to LR_GWESExplorer/snps.aln file.}

\item{lr_pos_path}{Path to LR_GWESExplorer/snps.loci file.}

\item{sr_tophits_path}{Path to Tophits/sr_tophits.tsv file. If SR link tophits should be visualised, the two sr_files (below) must also be provided here.}

\item{sr_fasta_path}{Path to SR_GWESExplorer/snps.aln file.}

\item{sr_pos_path}{Path to SR_GWESExplorer/snps.loci file.}

\item{ntop_links}{Specify the number of links to plot from tophits file(s). ntop_links will be selected from each file if both LR and SR links should be visualised (default = 10).}

\item{from}{draw the SNPs that are within the region starting <from> and <to> below (default = NULL). SNPs that are also in links with SNPs from the chosen region will be plotted.}

\item{to}{If <from> is provided, <to> muts also be provided and <ntop_links> will be ignored (default = NULL).}

\item{offset_metadata}{specify the separation between tree and metadata layer. These parameters should be modified to tune the plot dimensions as required. (default = NULL)}

\item{offset_alleles}{specify the speciation between tree and alleles layer (default = NULL)}

\item{width_metadata}{specify the width of the metadata layer as a proportion of the tree size (default = NULL)}

\item{width_alleles}{specify the width of the alleles layer as a proportion of the tree size (default = NULL)}

\item{plot_save_path}{path to save the tree plot (default = NULL, show plot instead of saving)}

\item{plot_height}{if saving, specify the plot height in inches (default = 20)}

\item{plot_width}{if saving, specify the plot height in inches (default = 15). Modify these numbers to tune the plot dimensions as required.}
}
\description{
Function to generate the bird's-eye view of the genomewide LD structure (measured as weighted-MI)
}
\examples{
\dontrun{
genomewide_LDMap(lr_links_path, sr_links_path, plot_save_path)
}

}
